egui_file_dialog

Module config

Re-exports

pub use labels::FileDialogLabels;

pub use keybindings::FileDialogKeyBindings;

pub use keybindings::KeyBinding;

Modules

keybindings 🔒

labels 🔒

Structs

FileDialogConfig

Contains configuration values of a file dialog.

FileFilter

Defines a specific file filter that the user can select from a dropdown.

Filter

Function that returns true if the specific item matches the filter.

IconFilter

Sets a specific icon for directory entries.

PinnedFolder

Folder that the user pinned to the left sidebar.

QuickAccess

Stores a custom quick access section of the file dialog.

QuickAccessPath

Stores the display name and the actual path of a quick access link.

SaveExtension

Defines a specific file extension that the user can select when saving a file.

Enums

OpeningMode

Sets which directory is loaded when opening the file dialog.