egui_file_dialog::create_directory_dialog

Struct CreateDirectoryDialog

pub struct CreateDirectoryDialog {
    open: bool,
    init: bool,
    directory: Option<PathBuf>,
    input: String,
    error: Option<String>,
    scroll_to_error: bool,
    request_focus: bool,
    file_system: Arc<dyn FileSystem + Send + Sync>,
}

A dialog to create new folder.

Fields
open: bool

If the dialog is currently open

init: bool

If the update method is called for the first time. Used to initialize some stuff and scroll to the dialog.

directory: Option<PathBuf>

The directory that is currently open and where the folder is created.

input: String

Buffer to hold the data of the folder name input

error: Option<String>

This contains the error message if the folder name is invalid

scroll_to_error: bool

If we should scroll to the error in the next frame

request_focus: bool

If the text input should request focus in the next frame

file_system: Arc<dyn FileSystem + Send + Sync>
Implementations

impl CreateDirectoryDialog

pub fn from_filesystem(file_system: Arc<dyn FileSystem + Send + Sync>) -> Self

Creates a new dialog with default values

pub fn open(&mut self, directory: PathBuf)

Resets the dialog and opens it.

pub fn close(&mut self)

Closes and resets the dialog without creating the directory.

pub fn submit(&mut self) -> CreateDirectoryResponse

Tries to create the given folder.

pub fn update( &mut self, ui: &mut Ui, config: &FileDialogConfig, ) -> CreateDirectoryResponse

Main update function of the dialog. Should be called in every frame in which the dialog is to be displayed.

pub const fn is_open(&self) -> bool

Returns if the dialog is currently open

fn create_directory(&mut self) -> CreateDirectoryResponse

Creates a new folder in the current directory. The variable input is used as the folder name. Might change the error variable when an error occurred creating the new folder.

fn validate_input(&mut self, labels: &FileDialogLabels) -> Option<String>

Validates the folder name input. Returns None if the name is valid. Otherwise returns the error message.

fn create_error(&mut self, error: &str) -> String

Creates the specified error and sets to scroll to the error in the next frame.

fn reset(&mut self)

Resets the dialog. Configuration variables are not changed.