egui_file_dialog::data::directory_content

Struct DirectoryContent

pub struct DirectoryContent {
    state: DirectoryContentState,
    content: Vec<DirectoryEntry>,
    content_recv: Option<Arc<Mutex<Receiver<Result<Vec<DirectoryEntry>, Error>>>>>,
}

Contains the content of a directory.

Fields
state: DirectoryContentState

Current state of the directory content.

content: Vec<DirectoryEntry>

The loaded directory contents.

content_recv: Option<Arc<Mutex<Receiver<Result<Vec<DirectoryEntry>, Error>>>>>

Receiver when the content is loaded on a different thread.

Implementations

impl DirectoryContent

pub fn from_path( config: &FileDialogConfig, path: &Path, file_system: Arc<dyn FileSystem + Sync + Send + 'static>, filter: DirectoryFilter, ) -> Self

Create a new DirectoryContent object and loads the contents of the given path. Use include_files to include or exclude files in the content list.

fn with_thread( config: &FileDialogConfig, path: &Path, file_system: Arc<dyn FileSystem + Send + Sync + 'static>, filter: DirectoryFilter, ) -> Self

fn without_thread( config: &FileDialogConfig, path: &Path, file_system: &dyn FileSystem, filter: &DirectoryFilter, ) -> Self

pub fn update(&mut self) -> &DirectoryContentState

fn update_pending_state(&mut self) -> &DirectoryContentState

pub fn iter_range_mut( &mut self, range: Range<usize>, ) -> impl Iterator<Item = &mut DirectoryEntry>

Returns an iterator in the given range of the directory contents. No filters are applied using this iterator.

pub fn filtered_iter<'s>( &'s self, search_value: &'s str, ) -> impl Iterator<Item = &'s DirectoryEntry> + 's

pub fn filtered_iter_mut<'s>( &'s mut self, search_value: &'s str, ) -> impl Iterator<Item = &'s mut DirectoryEntry> + 's

pub fn reset_multi_selection(&mut self)

Marks each element in the content as unselected.

pub fn len(&self) -> usize

Returns the number of elements inside the directory.

pub fn push(&mut self, item: DirectoryEntry)

Pushes a new item to the content.