egui_file_dialog::data::directory_content

Struct DirectoryEntry

pub struct DirectoryEntry {
    path: PathBuf,
    metadata: Metadata,
    is_directory: bool,
    is_system_file: bool,
    is_hidden: bool,
    icon: String,
    pub selected: bool,
}

Contains the information of a directory item.

This struct is mainly there so that the information and metadata can be loaded once and not that a request has to be sent to the OS every frame using, for example, path.is_file().

Fields
path: PathBufmetadata: Metadatais_directory: boolis_system_file: boolis_hidden: boolicon: Stringselected: bool

If the item is marked as selected as part of a multi selection.

Implementations

impl DirectoryEntry

pub fn from_path( config: &FileDialogConfig, path: &Path, file_system: &dyn FileSystem, ) -> Self

Creates a new directory entry from a path

pub const fn metadata(&self) -> &Metadata

Returns the metadata of the directory entry.

pub fn path_eq(&self, other: &Self) -> bool

Checks if the path of the current directory entry matches the other directory entry.

pub const fn is_dir(&self) -> bool

Returns true if the item is a directory. False is returned if the item is a file or the path did not exist when the DirectoryEntry object was created.

pub const fn is_file(&self) -> bool

Returns true if the item is a file. False is returned if the item is a directory or the path did not exist when the DirectoryEntry object was created.

pub const fn is_system_file(&self) -> bool

Returns true if the item is a system file.

pub fn icon(&self) -> &str

Returns the icon of the directory item.

pub fn as_path(&self) -> &Path

Returns the path of the directory item.

pub fn to_path_buf(&self) -> PathBuf

Clones the path of the directory item.

pub fn file_name(&self) -> &str

Returns the file name of the directory item.

pub const fn is_hidden(&self) -> bool

Returns whether the path this DirectoryEntry points to is considered hidden.