egui_file_dialog

Module file_dialog

Structs

FileDialog

Represents a file dialog instance.

FileDialogStorage

Contains data of the FileDialog that should be stored persistently.

Enums

DialogMode

Represents the mode the file dialog is currently in.

DialogState

Represents the state the file dialog is currently in.

Type Aliases

FileDialogUiCallback 🔒

Callback type to inject a custom egui ui inside the file dialog’s ui.