egui_file_dialog::modals::overwrite_file_modal

Struct OverwriteFileModal

pub struct OverwriteFileModal {
    state: ModalState,
    path: PathBuf,
}

The modal that is used to ask the user if the selected path should be overwritten.

Fields
state: ModalState

The current state of the modal.

path: PathBuf

The path selected for overwriting.

Implementations

impl OverwriteFileModal

pub const fn new(path: PathBuf) -> Self

Creates a new modal object.

Arguments

path - The path selected for overwriting.

impl OverwriteFileModal

fn submit(&mut self)

Submits the modal and triggers the action to save the file.

fn cancel(&mut self)

Closes the modal without overwriting the file.