egui_file_dialog

Struct FileDialogStorage

pub struct FileDialogStorage {
    pub pinned_folders: Vec<PinnedFolder>,
    pub show_hidden: bool,
    pub show_system_files: bool,
    pub last_visited_dir: Option<PathBuf>,
    pub last_picked_dir: Option<PathBuf>,
}

Contains data of the FileDialog that should be stored persistently.

Fields
pinned_folders: Vec<PinnedFolder>

The folders the user pinned to the left sidebar.

show_hidden: bool

If hidden files and folders should be listed inside the directory view.

show_system_files: bool

If system files should be listed inside the directory view.

last_visited_dir: Option<PathBuf>

The last directory the user visited.

last_picked_dir: Option<PathBuf>

The last directory from which the user picked an item.